fix: some more cleanup, one tool per purpose
