+++ src/aibtc-cohort-0/contract-tools/generate-dao-contracts.ts
@@ -1,5 +1,4 @@
 import { ContractApiClient } from "../api/client";
-import { CONFIG, deriveChildAccount } from "../../utilities";
 
 export interface DeployDaoParams {
   tokenSymbol: string;
@@ -44,44 +43,4 @@ export async function generateDaoContracts(
     console.error("Error generating DAO contracts:", error);
     throw error;
   }
-}
-
-export async function deployDaoContracts(params: DeployDaoParams) {
-  const {
-    tokenSymbol,
-    tokenName,
-    tokenMaxSupply,
-    tokenUri,
-    logoUrl,
-    originAddress,
-    daoManifest,
-    tweetOrigin,
-    daoManifestInscriptionId,
-    network = CONFIG.NETWORK,
-  } = params;
-
-  // Generate contracts using the API
-  const apiClient = new ContractApiClient();
-  const customReplacements = {
-    token_symbol: tokenSymbol,
-    token_name: tokenName,
-    token_max_supply: tokenMaxSupply.toString(),
-    token_uri: tokenUri,
-    logo_url: logoUrl,
-    origin_address: originAddress,
-    dao_manifest: daoManifest,
-    tweet_origin: tweetOrigin,
-    dao_manifest_inscription_id: daoManifestInscriptionId || "",
-  };
-
-  // Generate all DAO contracts
-  const generatedContracts = await apiClient.generateDaoContracts(
-    network,
-    tokenSymbol,
-    customReplacements
-  );
-
-  // Here you would typically deploy these contracts
-  // For now, we'll just return the generated contracts
-  return generatedContracts;
 }
